scraper/url/requirements.txt
selenium
webdriver-manager
beautifulsoup4
undetected-chromedriver
Django>=4.2
psycopg2-binary
